+++ module/data/item-burden.mjs
@@ -12,17 +12,16 @@ export default class SdmBurden extends SdmItemBase {
       required: false,
       blank: true,
       initial: '',
-      choices: Object.entries(CONFIG.SDM.burdenType).reduce((acc, [key, value]) => {
+      choices: Object.entries({
+        ...CONFIG.SDM.burdenType,
+        '': 'TYPE.Burden'
+      }).reduce((acc, [key, value]) => {
         acc[key] = game.i18n.localize(value);
         return acc;
       }, {})
     });
 
-    schema.power = new fields.EmbeddedDataField(PowerDataModel, {
-      required: false,
-      nullable: true,
-      initial: null
-    });
+    schema.power = new fields.EmbeddedDataField(PowerDataModel);
 
     schema.cure_steps = new fields.SchemaField(
       {